backing into gas.properties didn't work
