+++ particula/next/gas/__init__.py
@@ -1,0 +1,29 @@
+"""Import all the gas modules, so they can be accessed from
+particula.next import gas
+"""
+
+# pylint: disable=unused-import
+# flake8: noqa
+# pyright: basic
+
+# from particula.next.gas.vapor_pressure_strategies import (
+#     ConstantVaporPressureStrategy,
+#     AntoineVaporPressureStrategy,
+#     ClausiusClapeyronStrategy,
+#     WaterBuckStrategy,
+# )
+# from particula.next.gas.vapor_pressure_builders import (
+#     ConstantBuilder,
+#     AntoineBuilder,
+#     ClausiusClapeyronBuilder,
+#     WaterBuckBuilder,
+# )
+# from particula.next.gas.vapor_pressure_factories import VaporPressureFactory
+# from particula.next.gas.species import GasSpecies
+# from particula.next.gas.species_builders import (
+#     GasSpeciesBuilder, PresetGasSpeciesBuilder,
+# )
+# from particula.next.gas.species_factories import GasSpeciesFactory
+# from particula.next.gas.atmosphere import Atmosphere
+# from particula.next.gas.atmosphere_builders import AtmosphereBuilder
+# from particula.next.gas import properties
