+++ particula/next/gas/properties/__init__.py
@@ -6,6 +6,18 @@ particula.next.gas.properties.
 # flake8: noqa
 # pyright: basic
 
-from particula.next.gas.properties.mean_free_path import molecule_mean_free_path
-from particula.next.gas.properties.dynamic_viscosity import get_dynamic_viscosity
-from particula.next.gas.properties.thermal_conductivity import get_thermal_conductivity
+from particula.next.gas.properties.mean_free_path import (
+    molecule_mean_free_path,
+)
+from particula.next.gas.properties.dynamic_viscosity import (
+    get_dynamic_viscosity,
+)
+from particula.next.gas.properties.thermal_conductivity import (
+    get_thermal_conductivity,
+)
+from particula.next.gas.properties.concentration_function import (
+    calculate_concentration,
+)
+from particula.next.gas.properties.partial_pressure_function import (
+    calculate_partial_pressure,
+)
