+++ particula/next/gas/tests/vapor_pressure_strategies_test.py
@@ -4,69 +4,11 @@ import numpy as np
 import pytest
 from particula.constants import GAS_CONSTANT
 from particula.next.gas.vapor_pressure_strategies import (
-    calculate_partial_pressure,
-    calculate_concentration,
     ConstantVaporPressureStrategy,
     AntoineVaporPressureStrategy,
     ClausiusClapeyronStrategy,
     WaterBuckStrategy,
 )
-
-
-def test_calculate_partial_pressure_scalar():
-    """Test the calculate_partial_pressure function with scalar values."""
-    # Test the function with scalar values
-    concentration = 1.0  # kg/m^3
-    molar_mass = 0.029  # kg/mol (approx. for air)
-    temperature = 298  # K (approx. 25 degrees Celsius)
-    expected_pressure = (concentration * GAS_CONSTANT.m *
-                         temperature) / molar_mass
-    assert calculate_partial_pressure(
-        concentration,
-        molar_mass,
-        temperature) == pytest.approx(expected_pressure)
-
-
-def test_calculate_partial_pressure_array():
-    """Test the calculate_partial_pressure function with NumPy arrays."""
-    # Test the function with NumPy arrays
-    concentration = np.array([1.0, 2.0])  # kg/m^3
-    molar_mass = np.array([0.029, 0.032])  # kg/mol (approx. for air and O2)
-    temperature = np.array([298, 310])  # K
-    expected_pressure = (concentration * GAS_CONSTANT.m *
-                         temperature) / molar_mass
-    np.testing.assert_array_almost_equal(
-        calculate_partial_pressure(concentration, molar_mass, temperature),
-        expected_pressure
-    )
-
-
-def test_calculate_concentration_scalar():
-    """Test the calculate_concentration function with scalar values."""
-    # Test the function with scalar values
-    partial_pressure = 101325  # Pa
-    molar_mass = 0.029  # kg/mol (approx. for air)
-    temperature = 298  # K (approx. 25 degrees Celsius)
-    expected_concentration = (partial_pressure * molar_mass) / \
-        (GAS_CONSTANT.m * temperature)
-    assert calculate_concentration(
-        partial_pressure,
-        molar_mass,
-        temperature) == pytest.approx(expected_concentration)
-
-
-def test_calculate_concentration_array():
-    """Test the calculate_concentration function with NumPy arrays."""
-    # Test the function with NumPy arrays
-    partial_pressure = np.array([101325, 202650])  # Pa
-    molar_mass = np.array([0.029, 0.032])  # kg/mol (approx. for air and O2)
-    temperature = np.array([298, 310])  # K
-    expected_concentration = (partial_pressure * molar_mass) / \
-        (GAS_CONSTANT.m * temperature)
-    np.testing.assert_array_almost_equal(
-        calculate_concentration(partial_pressure, molar_mass, temperature),
-        expected_concentration
-    )
 
 
 def test_constant_vapor_pressure_strategy():
