+++ particula/next/gas/vapor_pressure_strategies.py
@@ -12,52 +12,10 @@ from typing import Union
 from numpy.typing import NDArray
 import numpy as np
 
-from particula.constants import GAS_CONSTANT  # pyright: ignore
-
-
-def calculate_partial_pressure(
-    concentration: Union[float, NDArray[np.float64]],
-    molar_mass: Union[float, NDArray[np.float64]],
-    temperature: Union[float, NDArray[np.float64]],
-) -> Union[float, NDArray[np.float64]]:
-    """
-    Calculate the partial pressure of a gas from its concentration, molar mass,
-    and temperature.
-
-    Parameters:
-    - concentration (float): Concentration of the gas in kg/m^3.
-    - molar_mass (float): Molar mass of the gas in kg/mol.
-    - temperature (float): Temperature in Kelvin.
-
-    Returns:
-    - float: Partial pressure of the gas in Pascals (Pa).
-    """
-    # Calculate the partial pressure
-    return (concentration * float(GAS_CONSTANT.m) * temperature) / molar_mass
-
-
-def calculate_concentration(
-    partial_pressure: Union[float, NDArray[np.float64]],
-    molar_mass: Union[float, NDArray[np.float64]],
-    temperature: Union[float, NDArray[np.float64]],
-) -> Union[float, NDArray[np.float64]]:
-    """
-    Calculate the concentration of a gas from its partial pressure, molar mass,
-    and temperature using the ideal gas law.
-
-    Parameters:
-    - pressure (float or NDArray[np.float64]): Partial pressure of the gas
-    in Pascals (Pa).
-    - molar_mass (float or NDArray[np.float64]): Molar mass of the gas in kg/mol
-    - temperature (float or NDArray[np.float64]): Temperature in Kelvin.
-
-    Returns:
-    - concentration (float or NDArray[np.float64]): Concentration of the gas
-    in kg/m^3.
-    """
-    return (partial_pressure * molar_mass) / (
-        float(GAS_CONSTANT.m) * temperature
-    )
+from particula.next.gas.properties.concentration_function import (
+    calculate_concentration
+)
+from particula.constants import GAS_CONSTANT
 
 
 class VaporPressureStrategy(ABC):
